+++ docRoot/css/style.css
@@ -1,0 +1,73 @@
+body {
+  margin : 0;
+  background-color : #fff;
+}
+
+.grid-container {
+  width : 90%;
+  height : 500px;
+  margin : 0 auto;
+  display : -ms-grid;
+  display :     grid;
+  -ms-grid-columns : 250px 10px 1fr 10px 1fr 10px 200px 10px 1fr;
+  grid-template-columns : 250px 1fr 1fr 200px 1fr;
+  -ms-grid-rows : 50px 10px 1fr 10px 1fr;
+  grid-template-rows : 50px 1fr 1fr;
+  grid-template-areas : "area1 area1 area2 area2 area7" "area3 area4 area4 area5 area7" "area3 area6 area6 area6 area7";
+  grid-gap : 10px;
+  gap : 10px;
+}
+
+.grid-container > div {
+  border : solid #000 1px;
+}
+
+.area1 {
+  -ms-grid-row : 1;
+  -ms-grid-column : 1;
+  -ms-grid-column-span : 3;
+  grid-area : area1;
+}
+
+.area2 {
+  -ms-grid-row : 1;
+  -ms-grid-column : 5;
+  -ms-grid-column-span : 3;
+  grid-area : area2;
+}
+
+.area3 {
+  -ms-grid-row : 3;
+  -ms-grid-row-span : 3;
+  -ms-grid-column : 1;
+  grid-area : area3;
+}
+
+.area4 {
+  -ms-grid-row : 3;
+  -ms-grid-column : 3;
+  -ms-grid-column-span : 3;
+  grid-area : area4;
+}
+
+.area5 {
+  -ms-grid-row : 3;
+  -ms-grid-column : 7;
+  grid-area : area5;
+}
+
+.area6 {
+  -ms-grid-row : 5;
+  -ms-grid-column : 3;
+  -ms-grid-column-span : 5;
+  grid-area : area6;
+}
+
+.area7 {
+  -ms-grid-row : 1;
+  -ms-grid-row-span : 5;
+  -ms-grid-column : 9;
+  grid-area : area7;
+}
+
+/*# sourceMappingURL=style.css.map */
