+++ src/index.ejs
@@ -10,7 +10,7 @@
 </head>
 
 <body>
-  <h1>CSS Gridサンプル</h1>
+  <h1>CSS Gridサンプルss</h1>
   <div class="grid-container">
     <div class="area1">1</div>
     <div class="area2">2</div>
