+++ src/js/_a.js
@@ -1,0 +1,1 @@
+export const hoge = 'hoge';
