--- src/js/index2.js
@@ -1,4 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var _a_1 = require("./_a");
-console.log(" index2 hoge is " + _a_1.hoge);
