+++ tasks/config.js
@@ -1,0 +1,19 @@
+export const isProduction = process.env.NODE_ENV === 'production';
+
+// 入出力パス
+export const PATHS = {
+  src: './src/',
+  docRoot: './docRoot/',
+  dest: './build/',
+};
+
+export const scripts = {
+  srcRoot: `${PATHS.src}js`,
+  src: `${PATHS.src}js/**/*.js`,
+  dest: `${PATHS.dest}js`,
+};
+
+// export const sass = {
+//   src: `${PATHS.src}/scss/**/*.scss`,
+//   dest: `${PATHS.src}/css`,
+// };
